crm_project/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
sqlalchemy==2.0.30
bcrypt==4.1.3
pandas==2.2.2
openpyxl==3.1.3
reportlab==4.2.0
schedule==1.2.2
pytest==8.2.0
python-dotenv==1.0.1
